added admins only routes
